[lp-1810] : move Address class to validator folder
